requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
streamlit
python-dotenv
requests
pydantic > requirements.txt